+++ packages/components/src/SideDrawer/SideDrawer.css.d.ts
@@ -1,0 +1,14 @@
+declare const styles: {
+  readonly "drawer": string;
+  readonly "container": string;
+  readonly "subtle": string;
+  readonly "header": string;
+  readonly "footer": string;
+  readonly "hasShadow": string;
+  readonly "headerActions": string;
+  readonly "hideWhenEmpty": string;
+  readonly "overlay": string;
+  readonly "content": string;
+};
+export = styles;
+